requirements.txt
google-generativeai==0.3.0
youtube-transcript-api==1.2.1
python-dotenv==1.0.0
streamlit==1.30.0
